# specs/001-media-influence-detection/contracts/api.yaml
openapi: 3.1.0
info:
  title: MITDS API
  description: Media Influence Topology & Detection System REST API
  version: 1.0.0
  contact:
    name: MITDS Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.mitds.example/v1
    description: Production
...
paths:
...
  /jobs/{id}:
    get:
      summary: Get async job status
      tags: [Ingestion, Detection, Reports]
      operationId: getJobStatus
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatus'
components:
  schemas:
...
    JobStatus:
      type: object
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, completed, failed]
        progress:
          type: number
          minimum: 0
          maximum: 100
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        result_url:
          type: string
          format: uri
        error:
          type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []